% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_binary.R
\name{BJSM_binary}
\alias{BJSM_binary}
\title{BJSM binary}
\usage{
BJSM_binary(
  data,
  NUM_ARMS,
  pi_prior_dist,
  pi_prior.a,
  pi_prior.b,
  beta0_prior_dist,
  beta0_prior.a,
  beta0_prior.b,
  beta1_prior_dist,
  beta1_prior.a,
  beta1_prior.c,
  n_MCMC_chain,
  BURN.IN,
  MCMC_SAMPLE,
  ci = 0.95,
  six = TRUE,
  DTR = TRUE
)
}
\arguments{
\item{data}{trial data generated through function \code{trial_dataset}}

\item{NUM_ARMS}{number of treatment arms}

\item{pi_prior_dist}{prior distribution for \code{pi}, user can choose from "gamma", "beta", "pareto"}

\item{pi_prior.a}{parameter \code{a} of the prior distribution for \code{pi_1K}, a vector with three values, one for each treatment. Please check the \code{Details} section for more explaination}

\item{pi_prior.b}{parameter \code{b} of the prior distribution  for \code{pi_1K}, a vector with three values, one for each treatment. Please check the \code{Details} section for more explaination}

\item{beta0_prior_dist}{prior distribution for \code{beta0}, user can choose from "gamma", "beta", "pareto"}

\item{beta0_prior.a}{parameter \code{a} of the prior distribution for linkage parameter \code{beta0}. Please check the \code{Details} section for more explaination}

\item{beta0_prior.b}{parameter \code{b} of the prior distribution  for linkage parameter \code{beta0}. Please check the \code{Details} section for more explaination}

\item{beta1_prior_dist}{prior distribution for \code{beta1}, user can choose from "gamma", "beta", "pareto"}

\item{beta1_prior.a}{parameter \code{a} of the prior distribution for linkage parameter \code{beta1}. Please check the \code{Details} section for more explaination}

\item{beta1_prior.c}{parameter \code{b} of the prior distribution for linkage parameter \code{beta1}. Please check the \code{Details} section for more explaination}

\item{n_MCMC_chain}{number of MCMC chains, default to 1.}

\item{BURN.IN}{number of burn-in iterations for MCMC}

\item{MCMC_SAMPLE}{number of iterations for MCMC}

\item{ci}{coverage probability for credible intervals, default = 0.95}

\item{six}{TRUE or FALSE. If TRUE, will run the six beta model, if FALSE will run the two beta model. default = TRUE}

\item{DTR}{TRUE or FALSE. If TRUE, will also return the expected response rate of dynamic treatment regimens. default = TRUE}

\item{posterior_sample}{posterior samples of the link parameters and response rates generated through the MCMC process
\itemize{

\strong{\code{pi_hat_bjsm}}: estimate of response rate/treatment effect \cr

\strong{\code{se_hat_bjsm}}: standard error of the response rate \cr

\strong{\code{ci_pi_A}}: x\% credible intervals for A \cr

\strong{\code{ci_pi_B}}: x\% credible intervals for B \cr

\strong{\code{ci_pi_C}}: x\% credible intervals for C \cr

\strong{\code{diff_AB}}: estimate of differences between treatments A and B \cr

\strong{\code{diff_BC}}: estimate of differences between treatments B and C \cr

\strong{\code{diff_AC}}: estimate of differences between treatments A and C \cr

\strong{\code{ci_diff_AB}}: x\% credible intervals for the differences between A and B \cr

\strong{\code{ci_diff_BC}}: x\% credible intervals for the differences between B and C \cr

\strong{\code{ci_diff_AC}}: x\% credible intervals for the differences between A and C \cr

\strong{\code{beta0_hat}}: linkage parameter \code{beta0} estimates \cr

\strong{\code{beta1_hat}}: linkage parameter \code{beta1} estimates \cr

\strong{\code{ci_beta0_hat}}: linkage parameter \code{beta0} credible interval \cr

\strong{\code{ci_beta1_hat}}: linkage parameter \code{beta1} credible interval \cr

\strong{\code{pi_DTR_est}}: expected response rate of dynamic treatment regimens (DTRs) \cr
}}
}
\value{

}
\description{
BJSM (Bayesian Joint Stage Modeling) method that borrows information across both stages to estimate the individual response rate of each treatment (different active treatments, non-responders re-randomized; binary outcome) based on trial dataset
generated by function  \code{trial_data}
}
\details{
For \code{gamma} distribution, \code{prior.a} is the shape parameter \code{r}, \code{prior.b} is the rate parameter \code{lambda}. For \code{beta} distribution, \code{prior.a} is the shape parameter \code{a}, \code{prior.b} is the shape parameter \code{b}.
For \code{pareto} distribution, \code{prior.a} is the scale parameter \code{alpha}, \code{prior.b} is the shape parameter \code{c} (see page 29 of the jags user manual version 3.4.0). link: \url{http://www.stats.ox.ac.uk/~nicholls/MScMCMC14/jags_user_manual.pdf}

The individual response rate is regarded as a permanent feature of the treatment. The second stage outcome is modeled conditionally on the first stage results linking the first and
second stage response probabilities through linkage parameters.
}
\examples{
mydata = trial_dataset(trtA_I = 9, trtB_I = 12, trtC_I = 9, respA_I = 3,
    respB_I = 3, respC_I = 5, trtAA_II = 2, trtBB_II = 2, trtCC_II = 4,
    respA_II = 2, respB_II = 2, respC_II = 4, trtAB_II = 3, trtAC_II = 2,
    trtBA_II = 3, trtBC_II = 2, trtCA_II = 1, trtCB_II = 2, respAB_II = 3,
    respAC_II = 1, respBA_II = 2, respBC_II = 2, respCA_II = 0, respCB_II = 0)

BJSM_result = BJSM_binary(data = mydata, NUM_ARMS = 3, pi_prior_dist = "beta",
    pi_prior.a = c(0.4, 0.4, 0.4), pi_prior.b = c(1.6, 1.6, 1.6),
    beta0_prior_dist = "beta", beta0_prior.a = 1.6, beta0_prior.b = 0.4,
    beta1_prior_dist = "pareto", beta1_prior.a = 3, beta1_prior.c = 1,
    n_MCMC_chain = 1, BURN.IN = 10000, MCMC_SAMPLE = 60000, ci = 0.95,
    six = TRUE, DTR = TRUE)

}
\references{
Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs).
Statistics in medicine, 37(26), pp.3723-3732.

Chao, Y.C., Trachtman, H., Gipson, D.S., Spino, C., Braun, T.M. and Kidwell, K.M., 2020. Dynamic treatment regimens in small n, sequential, multiple assignment,
randomized trials: An application in focal segmental glomerulosclerosis. Contemporary clinical trials, 92, p.105989.
}
