% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_dose_binary.R
\name{BJSM_binary_dose}
\alias{BJSM_binary_dose}
\alias{print.BJSM_dose_binary}
\title{BJSM binary (dose level snSMART design)}
\usage{
BJSM_binary_dose(
  data,
  prior_dist,
  pi_prior,
  normal.par,
  beta_prior,
  n_MCMC_chain,
  BURN.IN,
  MCMC_SAMPLE,
  ci = 0.95
)

\method{print}{BJSM_dose_binary}(object, digits = 5, ...)
}
\arguments{
\item{data}{trial data generated through function \code{\link{trial_dataset_dose}} or \code{\link{data_simulation_dose}}}

\item{prior_dist}{vector of two values ("prior distribution for \code{pi}", "prior distribution for \code{beta}"). User can choose from "gamma", "beta", "pareto". e.g. prior_dist = c("beta", "gamma")}

\item{pi_prior}{vector of two values (a, b). \code{a} is the parameter \code{a} of the prior distribution for \code{pi} (response rate) of placebo. \code{b} is the parameter \code{b} of the prior distribution for \code{pi} of placebo. Please check the \code{Details} section for more explanation}

\item{normal.par}{vector of two values (normal.mean, normal.var). our function assumes that the logarithm of treatment effect ratio follows a Gaussian prior distribution \eqn{N(\mu, \sigma^2)}, that is \eqn{log(\pi_L/\pi_P)~N(normal.mean, normal.var)}, and \eqn{log(\pi_H/\pi_P)~N(normal.mean, normal.var)}. \code{normal.mean} is the mean of this Gaussian prior. \code{normal.var} is the variance of this Gaussian prior distribution}

\item{beta_prior}{vector of two values (a, b). \code{a} is the parameter \code{a} of the prior distribution for linkage parameter \code{beta}. \code{b} is the parameter b of the prior distribution  for linkage parameter \code{beta}}

\item{n_MCMC_chain}{number of MCMC chains, default to 1}

\item{BURN.IN}{number of burn-in iterations for MCMC}

\item{MCMC_SAMPLE}{number of iterations for MCMC}

\item{ci}{coverage probability for credible intervals, default = 0.95}

\item{digits}{the number of significant digits to use when printing}
}
\value{
\describe{
\item{posterior_sample}{posterior samples of the link parameters and response rates generated through the MCMC process}
\item{pi_hat_bjsm}{estimate of response rate/treatment effect}
\item{se_hat_bjsm}{standard error of the estimated response rate}
\item{ci_pi_P, ci_pi_L, ci_pi_H}{x\% credible intervals for treatment P, L, H}
\item{diff_PL, diff_LH, diff_PH}{estimate of differences between treatment P and L, L and H, P and H}
\item{se_PL, se_LH, se_PH}{standard error of the estimated differences between treatments}
\item{ci_diff_PL, ci_diff_LH, ci_diff_PH}{x\% credible intervals for the differences between treatment P and L, L and H, P and H}
\item{beta_hat}{linkage parameter beta estimates}
\item{se_beta}{standard error of the estimated value of beta}
\item{ci_beta_hat}{linkage parameter beta estimates}}
}
\description{
BJSM (Bayesian Joint Stage Modeling) method that borrows information across both stages to estimate the individual response rate of each dose level (placebo, low, high does; binary outcome) based on trial dataset
generated by function \code{\link{trial_dataset_dose}} or \code{\link{data_simulation_dose}}
}
\details{
For gamma distribution, \code{prior.a} is the shape parameter \code{r}, \code{prior.b} is the rate parameter \code{lambda}. For beta distribution, \code{prior.a} is the shape parameter \code{a}, \code{prior.b} is the shape parameter \code{b}.
For pareto distribution, \code{prior.a} is the scale parameter \code{alpha}, \code{prior.b} is the shape parameter \code{c} (see page 29 of the jags user manual version 3.4.0). link: \url{http://www.stats.ox.ac.uk/~nicholls/MScMCMC14/jags_user_manual.pdf}

The individual response rate is regarded as a permanent feature of the treatment. The second stage outcome is modeled conditionally on the first stage results linking the first and
second stage response probabilities through linkage parameters. The stage 1 response rate for treatment \eqn{k} is denoted as \eqn{\pi_k}. The stage 2 response rate for stage 1 responders to
treatment \eqn{k} who receive treatment \eqn{k'} in stage 2 is equal to \eqn{\beta_{1k}\pi_k'}. For nonresponders to treatment \eqn{k} in stage 1 who receive treatment \eqn{k*} in stage 2, the stage
2 response rate is equal to \eqn{\beta_{0k}\pi_k*}

Please refer to the paper listed under \code{reference} section for dose level snSMART design and detailed definition of parameters.
}
\examples{
mydata = trial_dataset_dose(trt = c(30, 30, 30), resp = c(5, 10, 15), trt_rep_P = c(3, 2), trt_rep_L = c(5, 5),
     trt_rep_H = c(8, 7), resp_rep_P = c(1, 2), resp_rep_L = c(2, 3), resp_rep_H = c(4, 6), trt_nrep_P = c(10, 15),
     trt_nrep_L = c(10, 10), trt_nrep_H = 15, resp_nrep_P = c(7, 8), resp_nrep_L = c(7, 6), resp_nrep_H = 10)

BJSM_dose_result = BJSM_binary_dose(data = mydata, prior_dist = c("beta", "gamma"),
    pi_prior = c(3, 17), normal.par = c(0.2, 100), beta_prior = c(2, 2),
    n_MCMC_chain = 2, BURN.IN = 10000, MCMC_SAMPLE = 60000, ci = 0.95)

summary(BJSM_dose_result)

}
\references{
Fang, F., Hochstedler, K.A., Tamura, R.N., Braun, T.M. and Kidwell, K.M., 2021. Bayesian methods to compare dose levels with placebo in a small n,
sequential, multiple assignment, randomized trial. Statistics in Medicine, 40(4), pp.963-977.
}
\seealso{
\code{\link{data_simulation_dose}} \cr
\code{\link{trial_dataset_dose}} \cr
\code{\link{JSRM_binary_dose}}
}
