% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_simulation.R, R/c_trialDataGen.R
\name{c_trialDataGen}
\alias{c_trialDataGen}
\title{Data Simulation (continuous snSMART design)}
\usage{
c_trialDataGen(
  treatInfo,
  treatCors,
  n,
  stay.ethical = NULL,
  switch.safety = NULL,
  wideForm = TRUE
)

c_trialDataGen(
  treatInfo,
  treatCors,
  n,
  stay.ethical = NULL,
  switch.safety = NULL,
  wideForm = TRUE
)
}
\arguments{
\item{treatInfo}{Treatment distribution information. Each row represents one treatment, format: (mean of treatment effects, sd of treatment effects, c(priming effect on trt A,  trt B, trt C)), see example for detail}

\item{treatCors}{Treatment correlations matrix. Format example:
[[ (A,A) (A,B) (A,C)
(B,A) (B,B) (B,C)
(C,A) (C,B) (C,C) ]], (A,B) denotes the treatment correlation between first stage treatment A and second stage treatment B}

\item{n}{vector of number of patients on each treatment e.g. n <- c(100, 100, 100)}

\item{stay.ethical}{numerical value, if stage 1 outcome is bigger than \code{stay.ethical} value, patient has probability of 1 of staying on the same treatment in stage 2}

\item{switch.safety}{numerical value, if stage 1 outcome is smaller than \code{switch.safety} value, patient has probability of 0 of staying on the same treatment in stage 2}

\item{wideForm}{whether to output the simulated dataset in wide form}
}
\value{
The simulated dataset with 7 variables: patient id, treatment 1, stage 1 outcome, probability of stay on the same treatment, realization of staying flag (1 = stay on the same treatment, 0 = rerandomization in stage 2), treatment 2, stage 2 outcome

The simulated dataset with 7 variables: patient id, treatment 1, stage 1 outcome, probability of stay on the same treatment, realization of staying flag (1 = stay on the same treatment, 0 = rerandomization in stage 2), treatment 2, stage 2 outcome
}
\description{
simulate data for snSMART with continuous outcome (non-responders re-randomized)

simulate data for snSMART with continuous outcome (non-responders re-randomized)
}
\examples{
treat.a<-c(70, 15, c(0,0,0))
treat.b<-c(50, 10, c(0,0,0))
treat.c<-c(60, 12, c(0,0,0))
treatInfo = rbind(treat.a, treat.b, treat.c)

treatCors<-diag(.9, 3)

switch.safety<-NULL
stay.ethical<- NULL
na<-100
nb<-100
nc<-100
n<-c(na,nb,nc)
trialData = c_trialDataGen(treatInfo, treatCors, n)

treat.a<-c(70, 15, c(0,0,0))
treat.b<-c(50, 10, c(0,0,0))
treat.c<-c(60, 12, c(0,0,0))
treatInfo = rbind(treat.a, treat.b, treat.c)

treatCors<-diag(.9, 3)

switch.safety<-NULL
stay.ethical<- NULL
na<-100
nb<-100
nc<-100
n<-c(na,nb,nc)
trialData = c_trialDataGen(treatInfo, treatCors, n, wideForm = FALSE)

}
\references{
Hartman, H., Tamura, R.N., Schipper, M.J. and Kidwell, K.M., 2021. Design and analysis considerations for utilizing a mapping function in a small sample,
sequential, multiple assignment, randomized trials with continuous outcomes. Statistics in Medicine, 40(2), pp.312-326.

Hartman, H., Tamura, R.N., Schipper, M.J. and Kidwell, K.M., 2021. Design and analysis considerations for utilizing a mapping function in a small sample,
sequential, multiple assignment, randomized trials with continuous outcomes. Statistics in Medicine, 40(2), pp.312-326.
}
