% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_binary}
\alias{data_binary}
\title{Data Binary}
\description{
sample dataset of snSMART (3 active treatment) with binary outcomes
}
\examples{
mydata = data_binary
LPJSM_result = LPJSM_binary(data = mydata, six = TRUE, DTR = TRUE)

}
\keyword{data}
