% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_group_seq_1step.R
\name{data_gen_group_seq_1step}
\alias{data_gen_group_seq_1step}
\title{Data generation of a group sequential snSMART with a one-step rule}
\usage{
data_gen_group_seq_1step(
  pi_1A,
  pi_1B,
  pi_1C,
  discount_y,
  discount_n1,
  discount_n2,
  rate,
  n.month,
  n.update = 1,
  drop_threshold = 0.5,
  NUM_ARMS,
  trt.days = 180,
  trt.break = 1,
  pi_prior_dist,
  pi_prior.a,
  pi_prior.b,
  beta0_prior_dist,
  beta0_prior.a,
  beta0_prior.b,
  beta1_prior_dist,
  beta1_prior.a,
  beta1_prior.c,
  MCMC_SAMPLE,
  BURN.IN,
  n_MCMC_chain,
  ci = 0.95,
  DTR = TRUE
)
}
\arguments{
\item{pi_1A}{first stage response rate of A}

\item{pi_1B}{first stage response rate of B}

\item{pi_1C}{first stage response rate of C}

\item{discount_y}{linkage parameters, for responders to treatment k in the first stage who received treatment k again in the second stage, the second stage response rate is equal to \code{discount_y * pi_IK}}

\item{discount_n1}{linkage parameters, vector of three values: 1) nonresponders to A receive treatment B in second stage, 2) nonresponders to B receive treatment A in second stage, 3) nonresponders to C receive A in the second stage}

\item{discount_n2}{linkage parameters, vector of three values: 1) nonresponders to A receive treatment C in second stage, 2) nonresponders to B receive treatment C in second stage, 3) nonresponders to C receive B in the second stage}

\item{rate}{accrual rate. Rate equals to 5 if participants are enrolled in the study at the rate of five people per month}

\item{n.month}{number of months to enroll subjects}

\item{n.update}{number of updates during the study. Default is 1 (number of interim analysis? need to check with Kelley)}

\item{drop_threshold}{a number between 0 and 1, if the posteriors probability of a treatment having the smallest response rate is bigger than the \code{drop_threshold}, this treatment will be dropped at that update. Default value is 0.5, the number of dropping threshold should be the same as the number update in the study}

\item{NUM_ARMS}{number of treatment arms}

\item{trt.days}{subjects' responses are evaluated \code{trt.days} days after receiving the treatment, default 180}

\item{trt.break}{second stage trt is given \code{trt.break} day after the evaluation, default 1}

\item{pi_prior_dist}{prior distribution for pi, user can choose from gamma, beta, pareto}

\item{pi_prior.a}{parameter a of the prior distribution for pi_1K, a vector with three values, one for each treatment. Please check the \code{Details} section for more explaination}

\item{pi_prior.b}{parameter b of the prior distribution  for pi_1K, a vector with three values, one for each treatment. Please check the \code{Details} section for more explaination}

\item{beta0_prior_dist}{prior distribution for beta0, user can choose from gamma, beta, pareto}

\item{beta0_prior.a}{parameter a of the prior distribution for linkage parameter beta0}

\item{beta0_prior.b}{parameter b of the prior distribution  for linkage parameter beta0}

\item{beta1_prior_dist}{prior distribution for beta1, user can choose from gamma, beta, pareto}

\item{beta1_prior.a}{parameter a of the prior distribution for linkage parameter beta1}

\item{beta1_prior.c}{parameter b of the prior distribution for linkage parameter beta1}

\item{MCMC_SAMPLE}{number of iterations for MCMC}

\item{BURN.IN}{number of burn-in iterations for MCMC}

\item{n_MCMC_chain}{number of MCMC chains, default to 1. If this is set to a number more than 1}

\item{ci}{coverage probability for credible intervals, default = 0.95}

\item{DTR, }{if TRUE, will also return the expected response rate of dynamic treatment regimens. default = TRUE}
}
\value{
return the simulated dataset (8 columns: first treatment time, first response time, second treatment time, second response time, treatment arm for first treatment,
response for first treatment, treatment arm for second treatment, response for second treatment), randomization probabilities before each update (the number of rows
equals to the number of updates plus 1), removed arm, the round of the update that removal of treatment arm occurs; End of trial analysis outcome: (comments added next
to the return function, similar to BJSM_binary)
}
\description{
generate data for the Group Sequential snSMART, which is designed based on the standard design of snSMART (3 treatments, non-responders re-randomized; binary outcome)
Use BJSM to conduct interim and final analysis
}
\references{
Chao, Y.C., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2020. A Bayesian group sequential small n sequential multiple‐assignment randomized trial. Journal of the Royal Statistical Society: Series C (Applied Statistics), 69(3), pp.663-680.
}
