% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{data_simulation}
\alias{data_simulation}
\title{data_simulation}
\usage{
data_simulation(
  p_IA,
  p_IB,
  p_IC,
  discount_y0,
  discount_y1,
  p_1nA_2B,
  p_1nB_2A,
  p_1nC_2A,
  n_A,
  n_B,
  n_C
)
}
\arguments{
\item{p_IA}{first stage response rate of A}

\item{p_IB}{first stage response rate of B}

\item{p_IC}{first stage response rate of C}

\item{discount_y0}{linkage parameters, for nonresponders to treatment k in the first stage who received treatment k' in the second stage, the second stage response rate in the second stage is equal to discount_y0 * pi_IK'}

\item{discount_y1}{linkage parameters, the second stage response rate for first stage responders is equal to discount_y1 * p_Ik}

\item{p_1nA_2B}{probability for first stage non-responders to A being randomized to arm B in second stage}

\item{p_1nB_2A}{probability for first stage non-responders to B being randomized to arm A in second stage}

\item{p_1nC_2A}{probability for first stage non-responders to C being randomized to arm A in second stage}

\item{n_A}{first stage sample size of A}

\item{n_B}{first stage sample size of B}

\item{n_C}{first stage sample size of C}
}
\value{

}
\description{
simulate data for snSMART
}
\examples{

}
