% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_seq.R
\name{group_seq}
\alias{group_seq}
\title{BJSM Method for interim analysis and final analysis of group sequential trial design}
\usage{
group_seq(
  data,
  rule.type = 1,
  interim = TRUE,
  drop_threshold = 0.5,
  drop_threshold_large = 0.5,
  drop_threshold_small = 0.5,
  NUM_ARMS,
  pi_prior_dist,
  pi_prior.a,
  pi_prior.b,
  beta0_prior_dist,
  beta0_prior.a,
  beta0_prior.b,
  beta1_prior_dist,
  beta1_prior.a,
  beta1_prior.c,
  MCMC_SAMPLE,
  BURN.IN,
  n_MCMC_chain,
  ci = 0.95,
  DTR = TRUE
)
}
\arguments{
\item{data}{data should be in the same format as the simulated data generated by function \code{data_gen_group_seq_1step} and \code{data_gen_group_seq_2step}, with 8 columns:
\code{time.1st.trt} (first treatment starts time), \code{time.1st.resp} (first response time), \code{time.2nd.trt} (second treatment starts time), \code{time.2nd.resp} (second response time),
\code{trt.1st} (treatment arm for first treatment), \code{resp.1st} (response for first treatment), \code{trt.2nd} (treatment arm for second treatment), \code{resp.2nd} (response for second
treatment) data yet to be observed should be marked as "\code{NA}"}

\item{rule.type}{whether this trial is designed as a 1 step rule test  (\code{rule.type = 1}) or 2 step rule test (\code{rule.type = 2})}

\item{interim}{whether user is analyzing an interim analysis via BJSM (\code{interim} = TRUE) or an final analysis via BJSM (\code{interim} = FALSE)}

\item{drop_threshold}{only assign value to this parameter when \code{rule.type = 1} and \code{interim = TRUE}. a number between 0 and 1, if the posteriors probability of a treatment having the smallest response rate is bigger than the \code{drop_threshold}, this treatment will be dropped at that update. Default value is 0.5}

\item{drop_threshold_large}{only assign value to this parameter when \code{rule.type = 2} and \code{interim = TRUE}. a number between 0 and 1. Default value is 0.5}

\item{drop_threshold_small}{only assign value to this parameter when \code{rule.type = 2} and \code{interim = TRUE}. a number between 0 and 1. Default value is 0.5. See the details section for more explanation}

\item{NUM_ARMS}{number of treatment arms}

\item{pi_prior_dist}{prior distribution for \code{pi}, user can choose from "gamma", "beta", "pareto"}

\item{pi_prior.a}{parameter a of the prior distribution for \code{pi_1K}, a vector with three values, one for each treatment. Please check the \code{Details} section for more explanation}

\item{pi_prior.b}{parameter b of the prior distribution  for \code{pi_1K}, a vector with three values, one for each treatment. Please check the \code{Details} section for more explanation}

\item{beta0_prior_dist}{prior distribution for \code{beta0}, user can choose from "gamma", "beta", "pareto"}

\item{beta0_prior.a}{parameter a of the prior distribution for linkage parameter \code{beta0}. Please check the \code{Details} section for more explanation}

\item{beta0_prior.b}{parameter b of the prior distribution  for linkage parameter \code{beta0}. Please check the \code{Details} section for more explanation}

\item{beta1_prior_dist}{prior distribution for \code{beta1}, user can choose from "gamma", "beta", "pareto"}

\item{beta1_prior.a}{parameter a of the prior distribution for linkage parameter \code{beta1}. Please check the \code{Details} section for more explanation}

\item{beta1_prior.c}{parameter b of the prior distribution for linkage parameter \code{beta1}. Please check the \code{Details} section for more explanation}

\item{MCMC_SAMPLE}{number of iterations for MCMC}

\item{BURN.IN}{number of burn-in iterations for MCMC}

\item{n_MCMC_chain}{number of MCMC chains, default to 1}

\item{ci}{coverage probability for credible intervals, default = 0.95. only assign value to this parameter when \code{interim = FALSE}.}

\item{DTR, }{if TRUE, will also return the expected response rate of dynamic treatment regimens. default = TRUE. only assign value to this parameter when \code{interim = FALSE}.}
}
\value{
if \code{interim = TRUE}, this function returns either 0 - no arm is dropped, or A/B/C - arm A/B/C is dropped \cr

if \code{interim = FALSE}, this function returns: \cr

\strong{\code{posterior_sample}}: posterior samples of the link parameters and response rates generated through the MCMC process \cr

\strong{\code{pi_hat_bjsm}}: estimate of response rate/treatment effect \cr

\strong{\code{se_hat_bjsm}}: standard error of the response rate \cr

\strong{\code{ci_pi_A}}: x\% credible intervals for A \cr

\strong{\code{ci_pi_B}}: x\% credible intervals for B \cr

\strong{\code{ci_pi_C}}: x\% credible intervals for C \cr

\strong{\code{diff_AB}}: estimate of differences between treatments A and B \cr

\strong{\code{diff_BC}}: estimate of differences between treatments B and C \cr

\strong{\code{diff_AC}}: estimate of differences between treatments A and C \cr

\strong{\code{ci_diff_AB}}: x\% credible intervals for the differences between A and B \cr

\strong{\code{ci_diff_BC}}: x\% credible intervals for the differences between B and C \cr

\strong{\code{ci_diff_AC}}: x\% credible intervals for the differences between A and C \cr

\strong{\code{beta0_hat}}: linkage parameter \code{beta0} estimates \cr

\strong{\code{beta1_hat}}: linkage parameter \code{beta1} estimates \cr

\strong{\code{ci_beta0_hat}}: linkage parameter \code{beta0} credible interval \cr

\strong{\code{ci_beta1_hat}}: linkage parameter \code{beta1} credible interval \cr

\strong{\code{pi_DTR_est}}: expected response rate of dynamic treatment regimens (DTRs) \cr
}
\description{
After obtain real trial data, this function can be used to decide which arm to drop for an interim analysis or provide a full analysis for a final analysis
}
\details{
For \code{gamma} distribution, \code{prior.a} is the shape parameter \code{r}, \code{prior.b} is the rate parameter \code{lambda}. For \code{beta} distribution, \code{prior.a} is the shape parameter \code{a}, \code{prior.b} is the shape parameter \code{b}.
For \code{pareto} distribution, \code{prior.a} is the scale parameter \code{alpha}, \code{prior.b} is the shape parameter \code{c} (see page 29 of the jags user manual version 3.4.0). link: \url{http://www.stats.ox.ac.uk/~nicholls/MScMCMC14/jags_user_manual.pdf}

The individual response rate is regarded as a permanent feature of the treatment. The second stage outcome is modeled conditionally on the first stage results linking the first and
second stage response probabilities through linkage parameters.

(paper provided in the reference section, section 2.2.2 Bayesian decision rules. drop_threshold_large and drop_threshold_small are corresponding to \eqn{tau_l} and \eqn{phi_l} respectively)
}
\examples{
mydata = patient_entry

result1 = group_seq(data = mydata, rule.type = 1, interim = TRUE, drop_threshold=0.5, NUM_ARMS = 3, pi_prior_dist = "beta", pi_prior.a =  c(0.4,0.4,0.4), pi_prior.b = c(1.6, 1.6, 1.6), beta0_prior_dist = "beta",
beta0_prior.a = 1.6, beta0_prior.b = 0.4, beta1_prior_dist = "pareto", beta1_prior.a = 3, beta1_prior.c = 1, MCMC_SAMPLE = 60000, BURN.IN = 10000,
n_MCMC_chain = 1)

result2 = group_seq(data = mydata, rule.type = 2, interim = TRUE, drop_threshold_large=0.5, drop_threshold_small = 0.4, NUM_ARMS = 3, pi_prior_dist = "beta", pi_prior.a =  c(0.4,0.4,0.4), pi_prior.b = c(1.6, 1.6, 1.6), beta0_prior_dist = "beta",
beta0_prior.a = 1.6, beta0_prior.b = 0.4, beta1_prior_dist = "pareto", beta1_prior.a = 3, beta1_prior.c = 1, MCMC_SAMPLE = 60000, BURN.IN = 10000,
n_MCMC_chain = 1)

mydata = patient_entry_full
result3 = group_seq(data = mydata, rule.type = 2, interim = FALSE, NUM_ARMS = 3, pi_prior_dist = "beta", pi_prior.a =  c(0.4,0.4,0.4), pi_prior.b = c(1.6, 1.6, 1.6), beta0_prior_dist = "beta",
beta0_prior.a = 1.6, beta0_prior.b = 0.4, beta1_prior_dist = "pareto", beta1_prior.a = 3, beta1_prior.c = 1, MCMC_SAMPLE = 60000, BURN.IN = 10000,
n_MCMC_chain = 1, ci = 0.95, DTR = TRUE)

}
\references{
Chao, Y.C., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2020. A Bayesian group sequential small n sequential multiple‐assignment randomized trial. Journal of the Royal Statistical Society: Series C (Applied Statistics), 69(3), pp.663-680.
}
