% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSRM_dose_binary.R
\name{JSRM_binary_dose}
\alias{JSRM_binary_dose}
\alias{summary.JSRM_binary}
\alias{print.JSRM_binary}
\title{JSRM for dose level snSMART}
\usage{
JSRM_binary_dose(data)

\method{summary}{JSRM_binary}(object, digits = 5, ...)

\method{print}{JSRM_binary}(object, digits = 5, ...)
}
\arguments{
\item{data}{data format produced by the  \code{\link{trial_dataset_dose}} and \code{\link{data_simulation_dose}} function}

\item{digits}{the number of significant digits to use when printing}
}
\value{
a \code{list} containing
\itemize{
\item{\code{GEE_output}}{ - original output of the GEE (geeglm) model}
\item{\code{pi_hat}}{ - estimate of response rate/treatment effect}
\item{\code{sd_pi_hat}}{ - standard error of the response rate}
}
}
\description{
A joint-stage regression model (JSRM) is a frequentist modeling approach that incorporates the responses of both stages as repeated measurements for each subject.
Generalized estimating equations (GEE) are used to estimate the response rates of each dose level.
}
\examples{
mydata = trial_dataset_dose(trt = c(30, 30, 30), resp = c(5, 10, 15), trt_rep_P = c(3, 2), trt_rep_L = c(5, 5),
     trt_rep_H = c(8, 7), resp_rep_P = c(1, 2), resp_rep_L = c(2, 3), resp_rep_H = c(4, 6), trt_nrep_P = c(10, 15),
     trt_nrep_L = c(10, 10), trt_nrep_H = 15, resp_nrep_P = c(7, 8), resp_nrep_L = c(7, 6), resp_nrep_H = 10)

JSRM_result = JSRM_binary_dose(data = mydata)

summary(JSRM_result)

}
\references{
Fang, F., Hochstedler, K.A., Tamura, R.N., Braun, T.M. and Kidwell, K.M., 2021. Bayesian methods to compare dose levels with placebo in a small n,
sequential, multiple assignment, randomized trial. Statistics in Medicine, 40(4), pp.963-977.
}
\seealso{
\code{\link{data_simulation_dose}} \cr
\code{\link{trial_dataset_dose}} \cr
\code{\link{BJSM_binary_dose}}
}
