% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{sample_size}
\alias{sample_size}
\title{snSMART Sample Size Calculation}
\usage{
sample_size(
  piA,
  piB,
  piC,
  beta1,
  beta0,
  coverage,
  power,
  muA,
  muB,
  muC,
  nA,
  nB,
  nC
)
}
\arguments{
\item{piA}{the response rate (ranges from 0.01 to 0.99) for treatment A}

\item{piB}{the response rate (ranges from 0.01 to 0.99) for treatment B}

\item{piC}{the response rate (ranges from 0.01 to 0.99) for treatment C}

\item{beta1}{the linkage parameter (ranges from 1.00 to 1/largest response rate) for first stage responders. (A smller value leads to more conservative sample size calculation because two stages are less correlated)}

\item{beta0}{the linkage parameter (ranges from 0.01 to 0.99) for first stage non-responders. A larger value leads to a more conservative sample size calculation because two stages are less correlated}

\item{coverage}{the coverage rate (ranges from 0.01 to 0.99) for the posterior difference of top two treatments}

\item{power}{the probability (ranges from 0.01 to 0.99) for identify the best treatment}

\item{muA}{the prior mean (ranges from 0.01 to 0.99) for treatment A}

\item{muB}{the prior mean (ranges from 0.01 to 0.99) for treatment B}

\item{muC}{the prior mean (ranges from 0.01 to 0.99) for treatment C}

\item{nA}{the prior sample size (larger than 0) for treatment A}

\item{nB}{the prior sample size (larger than 0) for treatment B}

\item{nC}{the prior sample size (larger than 0) for treatment C}
}
\value{
the estimated sample size per arm for an snSMART
}
\description{
conduct Bayesian sample size calculation for an snSMART design to distinguish the best treatment from the second-best treatment using the Bayesian joint stage model
}
\references{
Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs).
Statistics in medicine, 37(26), pp.3723-3732.
}
