% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snSMART-package.R
\docType{package}
\name{snSMART-package}
\alias{snSMART}
\alias{snSMART-package}
\title{snSMART: R package for snSMART data simulation, sample size calculation and trial data analysis}
\description{
The aim of the snSMART R package is to consolidate data simulation, sample size calculation and analysis functions for several snSMART (small sample sequential, multiple assignment, randomized trial) designs under one library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sidiwang/snSMART}
  \item Report bugs at \url{https://github.com/sidiwang/snSMART/issues}
}

}
\author{
\strong{Maintainer}: Michael Kleinsasser \email{mkleinsa@umich.edu}

Authors:
\itemize{
  \item Sidi Wang \email{sidiwang@umich.edu}
  \item Kelley Kidwell \email{kidwell@umich.edu}
}

}
\keyword{internal}
