% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_binary.R
\name{trial_dataset}
\alias{trial_dataset}
\title{Trial dataset}
\usage{
trial_dataset(
  trtA_I,
  trtB_I,
  trtC_I,
  respA_I,
  respB_I,
  respC_I,
  trtAA_II,
  trtBB_II,
  trtCC_II,
  respA_II,
  respB_II,
  respC_II,
  trtAB_II,
  trtAC_II,
  trtBA_II,
  trtBC_II,
  trtCA_II,
  trtCB_II,
  respAB_II,
  respAC_II,
  respBA_II,
  respBC_II,
  respCA_II,
  respCB_II
)
}
\arguments{
\item{trtA_I}{number of people who receive A in stage 1}

\item{trtB_I}{number of people who receive B in stage 1}

\item{trtC_I}{number of people who receive C in stage 1}

\item{respA_I}{number of people who respond to A in stage 1, who receive same trt in stage 2}

\item{respB_I}{number of people who respond to B in stage 1, who receive same trt in stage 2}

\item{respC_I}{number of people who respond to C in stage 1, who receive same trt in stage 2}

\item{trtAA_II}{number of 1st stage responders to A who receive A in stage 2}

\item{trtBB_II}{number of 1st stage responders to B who receive B in stage 2}

\item{trtCC_II}{number of 1st stage responders to C who receive C in stage 2}

\item{respA_II}{number of 1st stage responders who respond to A again in stage 2}

\item{respB_II}{number of 1st stage responders who respond to B again in stage 2}

\item{respC_II}{number of 1st stage responders who respond to C again in stage 2}

\item{trtAB_II}{number of 1st stage non-responders to A who receive B in stage 2}

\item{trtAC_II}{number of 1st stage non-responders to A who receive C in stage 2}

\item{trtBA_II}{number of 1st stage non-responders to B who receive A in stage 2}

\item{trtBC_II}{number of 1st stage non-responders to B who receive C in stage 2}

\item{trtCA_II}{number of 1st stage non-responders to C who receive A in stage 2}

\item{trtCB_II}{number of 1st stage non-responders to C who receive B in stage 2}

\item{respAB_II}{number of 1st stage non-responders to A who respond to B in stage 2}

\item{respAC_II}{number of 1st stage non-responders to A who respond to C in stage 2}

\item{respBA_II}{number of 1st stage non-responders to B who respond to A in stage 2}

\item{respBC_II}{number of 1st stage non-responders to B who respond to C in stage 2}

\item{respCA_II}{number of 1st stage non-responders to C who respond to A in stage 2}

\item{respCB_II}{number of 1st stage non-responders to C who respond to B in stage 2}
}
\value{
a \code{matrix} with the following data sets row binded
\itemize{
\item{data_A.A.Y}{ - }
\item{data_A.B.Y}{ - }
\item{data_A.C.Y}{ - }
\item{data_A}{ - }
\item{data_A.A.}{ - }
\item{data_A.A.}{ - }
\item{data_A.A.}{ - }
\item{data_A.A.}{ - }
\item{data_A.A.}{ - }
\item{data_A.A.}{ - }
\item{data_A.A.}{ - }
\item{data_A.A.}{ - }
}
}
\description{
Generate trial dataset with or without missing value
}
\examples{


}
