% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_dose_binary.R
\name{trial_dataset_dose}
\alias{trial_dataset_dose}
\title{Trial dataset (dose level snSMART design)}
\usage{
trial_dataset_dose(
  trt,
  resp,
  trt_rep_P,
  trt_rep_L,
  trt_rep_H,
  resp_rep_P,
  resp_rep_L,
  resp_rep_H,
  trt_nrep_P,
  trt_nrep_L,
  trt_nrep_H,
  resp_nrep_P,
  resp_nrep_L,
  resp_nrep_H
)
}
\arguments{
\item{trt}{vector of 3 values - c(number of people who receive placebo in stage 1, number of people who receive low dose treatment in stage 1, number of people who receive high dose treatment in stage 1)}

\item{resp}{vector of 3 values - c(number of people who respond to placebo in stage 1, number of people who respond to low dose treatment in stage 1, number of people who respond to high dose treatment in stage 1)}

\item{trt_rep_P}{vector of 2 values - c(number of 1st stage responders to placebo who receive low dose treatment in stage 2, number of 1st stage responders to placebo who receive high dose treatment in stage 2)}

\item{trt_rep_L}{vector of 2 values - c(number of 1st stage responders to low dose treatment who receive low dose treatment in stage 2, number of 1st stage responders to low dose treatment who receive high dose treatment in stage 2)}

\item{trt_rep_H}{vector of 2 values - c(number of 1st stage responders to high dose treatment who receive low dose treatment in stage 2, number of 1st stage responders to high dose treatment who receive high dose treatment in stage 2)}

\item{resp_rep_P}{vector of 2 values - c(number of 1st stage responders to placebo who also respond to low dose treatment in stage 2, number of 1st stage responders to placebo who also respond to high dose treatment in stage 2)}

\item{resp_rep_L}{vector of 2 values - c(number of 1st stage responders to low dose treatment who also respond to low dose treatment  in stage 2, number of 1st stage responders to low dose treatment who also respond to high dose treatment in stage 2)}

\item{resp_rep_H}{vector of 2 values - c(number of 1st stage responders to high dose treatment who also respond to low dose treatment in stage 2, number of 1st stage responders to high dose treatment who also respond to high dose treatment in stage 2)}

\item{trt_nrep_P}{vector of 2 values - c(number of 1st stage non-responders to placebo who receive low dose treatment in stage 2, number of 1st stage non-responders to placebo who receive high dose treatment in stage 2)}

\item{trt_nrep_L}{vector of 2 values -c(number of 1st stage non-responders to low dose treatment who receive low dose treatment again in stage 2, number of 1st stage non-responders to low dose treatment who receive high dose treatment in stage 2)}

\item{trt_nrep_H}{number of 1st stage non-responders to high dose treatment who receive high dose treatment again in stage 2}

\item{resp_nrep_P}{vector of 2 values - c(number of 1st stage non-responders to placebo who respond to low dose treatment in stage 2, number of 1st stage non-responders to placebo who respond to high dose treatment in stage 2)}

\item{resp_nrep_L}{vector of 2 values - c(number of 1st stage non-responders to low dose treatment who respond to low dose treatment in stage 2, number of 1st stage non-responders to low dose treatment who respond to high dose treatment in stage 2)}

\item{resp_nrep_H}{number of 1st stage non-responders to high dose treatment who respond to high dose treatment in stage 2}
}
\value{
a \code{matrix} of the trial dataset with 4 columns: treatment_stageI, response_stageI, treatment_stageII, response_stageII
}
\description{
Generate trial dataset based on the exact number of responders/non-responders for dose level snSMART design (placebo, low, high dose; binary outcome) without missing value.
Useful for recording real snSMART trial result or generating specific simulation scenario. Not to be confused with \code{\link{data_simulation_dose}}.
}
\examples{
mydata = trial_dataset_dose(trt = c(30, 30, 30), resp = c(5, 10, 15), trt_rep_P = c(3, 2), trt_rep_L = c(5, 5),
     trt_rep_H = c(8, 7), resp_rep_P = c(1, 2), resp_rep_L = c(2, 3), resp_rep_H = c(4, 6), trt_nrep_P = c(10, 15),
     trt_nrep_L = c(10, 10), trt_nrep_H = 15, resp_nrep_P = c(7, 8), resp_nrep_L = c(7, 6), resp_nrep_H = 10)

}
\references{
Fang, F., Hochstedler, K.A., Tamura, R.N., Braun, T.M. and Kidwell, K.M., 2021. Bayesian methods to compare dose levels with placebo in a small n,
sequential, multiple assignment, randomized trial. Statistics in Medicine, 40(4), pp.963-977.
}
\seealso{
\code{\link{data_simulation_dose}} \cr
\code{\link{BJSM_binary_dose}} \cr
\code{\link{JSRM_binary_dose}}
}
