% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trialDataMF}
\alias{trialDataMF}
\title{Data Mapping Function}
\description{
sample dataset of snSMART (mapping function) with continuous outcomes
}
\examples{
trialData = trialDataMF

BJSM_result = BJSM_c(data = trialData, xi_prior.mean = c(50, 50, 50),
    xi_prior.sd = c(50, 50, 50), phi3_prior.sd = 20, n_MCMC_chain = 1,
    n.adapt = 1000, MCMC_SAMPLE = 5000, ci = 0.95, n.digits = 5)

summary(BJSM_result)
print(BJSM_result)

}
\keyword{data}
