% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_binary.R
\name{BJSM_binary}
\alias{BJSM_binary}
\title{BJSM binary}
\usage{
BJSM_binary(
  data,
  NUM_ARMS,
  pi_prior_dist,
  pi_prior.a,
  pi_prior.b,
  beta0_prior_dist,
  beta0_prior.a,
  beta0_prior.b,
  beta1_prior_dist,
  beta1_prior.a,
  beta1_prior.c,
  n_MCMC_chain,
  BURN.IN,
  MCMC_SAMPLE,
  ci = 0.95,
  six = TRUE
)
}
\arguments{
\item{NUM_ARMS}{number of treatment arms}

\item{pi_prior_dist}{prior distribution for pi, user can choose from gamma, beta, pareto}

\item{pi_prior.a}{alpha parameter of the prior beta distribution for pi_1K, a vector with three values, one for each treatment}

\item{pi_prior.b}{beta parameter of the prior beta distribution  for pi_1K, a vector with three values, one for each treatment}

\item{beta0_prior_dist}{prior distribution for beta0, user can choose from gamma, beta, pareto}

\item{beta0_prior.a}{alpha parameter of the prior beta distribution for linkage parameter beta0}

\item{beta0_prior.b}{beta parameter of the prior beta distribution  for linkage parameter beta0}

\item{beta1_prior_dist}{prior distribution for beta1, user can choose from gamma, beta, pareto}

\item{beta1_prior.a}{scale parameter of the prior pareto distribution for linkage parameter beta1}

\item{beta1_prior.c}{shape parameter of the prior pareto distribution for linkage parameter beta1}

\item{n_MCMC_chain}{number of MCMC chains, default to 1. If this is set to a number more than 1}

\item{BURN.IN}{number of burn-in iterations for MCMC}

\item{MCMC_SAMPLE}{number of iterations for MCMC}

\item{ci}{coverage probability for credible intervals}

\item{six, }{if TRUE, will run the six beta model, if FALSE will run the two beta model}
}
\value{

}
\description{
What does this function do?
}
\details{
For gamma, prior.a is r, prior.b is lambda. For beta, prior.a is alpha, prior.b is beta.
For pareto, prior.a is alpha, prior.b is c (see page 29 of the jags user manual version 3.4.0).
}
\examples{

}
