% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_dose_binary.R
\name{BJSM_binary_dose}
\alias{BJSM_binary_dose}
\title{BJSM binary (dose level design)}
\usage{
BJSM_binary_dose(
  data,
  NUM_ARMS,
  pi_prior_dist,
  pi_prior.a,
  pi_prior.b,
  normal.mean,
  normal.var,
  beta_prior_dist,
  beta_prior.a,
  beta_prior.b,
  n_MCMC_chain,
  BURN.IN,
  MCMC_SAMPLE,
  ci = 0.95
)
}
\arguments{
\item{data}{trial data generated through function \code{trial_dataset_dose}}

\item{NUM_ARMS}{number of treatment arms}

\item{pi_prior_dist}{prior distribution for pi (response rate) of placebo, user can choose from gamma, beta, pareto}

\item{pi_prior.a}{parameter a of the prior distribution for pi (response rate) of placebo. Please check the \code{Details} section for more explaination}

\item{pi_prior.b}{parameter b of the prior distribution for pi (response rate) of placebo. Please check the \code{Details} section for more explaination}

\item{normal.mean}{our function assumes that the logarithm of treatment effect ratio follows a Gaussian prior distribution \verb{N(\\mu, \\sigma^2)}, that is \verb{log(\\pi_L/\\pi_P)~N(normal.mean, normal.var}, and \verb{log(\\pi_H/\\pi_P)~N(normal.mean, normal.var}. \code{normal.mean} is the mean of this Gaussian prior}

\item{normal.var}{similar to above, \code{normal.var} is the variance of this Gaussian prior distribution}

\item{beta_prior_dist}{prior distribution for beta, user can choose from gamma, beta, pareto}

\item{beta_prior.a}{parameter a of the prior distribution for linkage parameter beta}

\item{beta_prior.b}{parameter b of the prior distribution  for linkage parameter beta}

\item{n_MCMC_chain}{number of MCMC chains, default to 1. If this is set to a number more than 1}

\item{BURN.IN}{number of burn-in iterations for MCMC}

\item{MCMC_SAMPLE}{number of iterations for MCMC}

\item{ci}{coverage probability for credible intervals, default = 0.95}
}
\value{
posterior_sample: posterior samples of the link parameters and response rates generated through the MCMC process
pi_hat_bjsm: estimate of response rate/treatment effect
se_hat_bjsm: standard error of the response rate
ci_pi_P: x\% credible intervals for A
ci_pi_L: x\% credible intervals for B
ci_pi_H: x\% credible intervals for C
diff_PL: estimate of differences between treatments A and B
diff_LH: estimate of differences between treatments B and C
diff_PH: estimate of differences between treatments A and C
ci_diff_PL: x\% credible intervals for the differences between A and B
ci_diff_LH: x\% credible intervals for the differences between B and C
ci_diff_PH: x\% credible intervals for the differences between A and C
beta_hat: linkage parameter beta estimates
ci_beta_hat: linkage parameter beta estimates
}
\description{
BJSM (Bayesian Joint Stage Modeling) method that borrows information across both stages to estimate the individual response rate of each treatment based on trial dataset
generated by function  \code{trial_data}
}
\details{
For gamma distribution, prior.a is the shape parameter r, prior.b is the rate parameter lambda. For beta distribution, prior.a is the shape parameter a, prior.b is the shape parameter b.
For pareto distribution, prior.a is the scale parameter alpha, prior.b is the shape parameter c (see page 29 of the jags user manual version 3.4.0). link: http://www.stats.ox.ac.uk/~nicholls/MScMCMC14/jags_user_manual.pdf

The individual response rate is regarded as a permanent feature of the treatment. The second stage outcome is modeled conditionally on the first stage results linking the first and
second stage response probabilities through linkage parameters.
}
\examples{
mydata = trial_dataset_dose(trtP_I = 30, trtL_I = 30, trtH_I = 30, respP_I = 5,
    respL_I = 10, respH_I = 15, trtPL_II = 3, trtPH_II = 2, trtLL_II = 5,
    trtLH_II = 5, trtHL_II = 8, trtHH_II = 7, respPL_II = 1, respPH_II = 2,
    respLL_II = 2, respLH_II = 3, respHL_II = 4, respHH_II = 6,
    trtNPL_II = 10, trtNPH_II = 15, trtNLH_II = 10, trtNLL_II = 10,
    trtNHH_II = 15, respNPL_II = 7, respNPH_II = 8, respNLH_II = 6,
    respNLL_II = 7, respNHH_II = 10)

BJSM_result = BJSM_binary(data = mydata, NUM_ARMS = 3, pi_prior_dist = "beta",
    pi_prior.a = 3, pi_prior.b = 17, beta_prior_dist = "gamma",
    normal.mean = 0.2, normal.var = 100, beta_prior.a = 2, beta_prior.b = 2,
    n_MCMC_chain = 2, BURN.IN = 10000, MCMC_SAMPLE = 60000, ci = 0.95)

}
\references{
Fang, F., Hochstedler, K.A., Tamura, R.N., Braun, T.M. and Kidwell, K.M., 2021. Bayesian methods to compare dose levels with placebo in a small n,
sequential, multiple assignment, randomized trial. Statistics in Medicine, 40(4), pp.963-977.
}
