% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_c.R
\name{BJSM_c}
\alias{BJSM_c}
\title{BJSM continuous (continuous outcome design)}
\usage{
BJSM_c(
  data,
  beta_prior.mean,
  beta_prior.sd,
  alpha3_prior.sd,
  n_MCMC_chain,
  n.adapt,
  MCMC_SAMPLE,
  ci = 0.95,
  n.digits
)
}
\arguments{
\item{data}{trial data generated through function \code{\link{c_trialDataGen}}}

\item{beta_prior.mean}{vector of mean of the prior distributions (normal distribution) for \code{beta} (treatment effect). Please check the \code{Details} section for more explaination}

\item{beta_prior.sd}{vector of standard deviation of the prior distributions (normal distribution) for \code{beta} (treatment effect). Please check the \code{Details} section for more explaination}

\item{alpha3_prior.sd}{standard deviation of the prior distribution (folded normal distribution) of \code{alpha3} (if the patient stays on the same treatment, \code{alpha3} is the cumulative effect of stage 1 that occurs on the treatment longer term). Please check the \code{Details} section for more explaination}

\item{n_MCMC_chain}{number of MCMC chains, default to 1}

\item{n.adapt}{the number of iterations for adaptation. If \code{n.adapt = 0} then no adaptation takes place}

\item{MCMC_SAMPLE}{number of iterations for MCMC}

\item{ci}{coverage probability for credible intervals, default = 0.95}

\item{n.digits}{number of digits to keep in the final output}
}
\value{
\describe{
\item{posterior_sample}{posterior samples of the link parameters and response rates generated through the MCMC process}
\item{mean_estimate}{BJSM estimate of each parameter:
\enumerate{
\item \code{alpha1} - lingering effect of the first treatment
\item \code{alpha3} - if the patient stays on the same treatment, \code{alpha3} is the cumulative effect of stage 1 that occurs on the treatment longer term
\item \code{betaj} - \code{betaj} parameters are the expected effect of treat j, j = A, B, C, in the first stage
\item \code{rho} is the inverse of the variance-covariance matrix of the multivariate distribution, first parameter indicates whether patient stayed on the same treatment (2) or not (1), second parameter
indicates the row number of the inverse of variance-covariance matrix, and the third parameter indicates the column number of the inverse of the variance-covariance matrix}
}
\item{ci_estimate}{x\% credible interval for each parameter. By default round to 2 decimal places, if more decimals are needed, please access the results by \verb{[YourResultName]$ci_estimates$CI_low} or \verb{[YourResultName]$ci_estimates$CI_high} }
}
}
\description{
BJSM (Bayesian Joint Stage Modeling) method that borrows information across both stages to estimate the individual response rate of each treatment (with continuous outcome) based on trial dataset
generated by function  \code{\link{c_trialDataGen}}
``
}
\details{
section 2.2.1 of the paper listed under \code{reference} provides a detailed description of the assumptions of the model.
}
\examples{
treat.a<-c(70, 15, c(0,0,0))
treat.b<-c(50, 10, c(0,0,0))
treat.c<-c(60, 12, c(0,0,0))
treatInfo = rbind(treat.a, treat.b, treat.c)

treatCors<-diag(.9, 3)

switch.safety<-NULL
stay.ethical<- NULL
na<-100
nb<-100
nc<-100
n<-c(na,nb,nc)
trialData = c_trialDataGen(treatInfo, treatCors, n)


BJSM_result = BJSM_c(data = trialData, beta_prior.mean = c(50, 50, 50),
    beta_prior.sd = c(50, 50, 50), alpha3_prior.sd = 20, n_MCMC_chain = 1,
    n.adapt = 1000, MCMC_SAMPLE = 5000, ci = 0.95, n.digits = 5)

}
\references{
Hartman, H., Tamura, R.N., Schipper, M.J. and Kidwell, K.M., 2021. Design and analysis considerations for utilizing a mapping function in a small sample,
sequential, multiple assignment, randomized trials with continuous outcomes. Statistics in Medicine, 40(2), pp.312-326.
}
\seealso{
\code{\link{c_trialSim}} \cr
\code{\link{c_trialDataGen}}
}
