% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_trialDataGen.R
\name{c_trialDataGen}
\alias{c_trialDataGen}
\title{Data Simulation (continuous snSMART design)}
\usage{
c_trialDataGen(
  stage1effects,
  stage2weights,
  stagecorrs,
  variance,
  n,
  stay.ethical = NULL,
  switch.safety = NULL,
  wideForm = TRUE
)
}
\arguments{
\item{stage1effects}{a vector of 3 values (\verb{mean effect of treatment A}, \verb{mean effect of treatment B}, \verb{main effect of treatment C}). e.g. \code{stage1effects = c(40, 50, 70)}}

\item{stage2weights}{a vector of 3 values (\verb{weight of stage I}, \verb{weight of stage II}, \verb{weight of staying on the same treatment}). The second stage mean outcome is a weighted average of the treatment effects from stage one and stage two with an additional effect if the patient stays on the same treatment. e.g. \code{stage2weights = c(0.2, 0.7, 0.3)}}

\item{stagecorrs}{a vector of 2 values (\verb{correlations between two stages if patient stays on the same treatment}, \verb{correlations between two stages if patent doesn't stay on the same treatment}). This allows those who stay on the same treatmetn to have a different correlation between stage one and stage two outcomes than those who switch treatments.}

\item{variance}{the variance of the treatment effects}

\item{n}{vector of number of patients on each treatment, e.g. \code{n <- c(100, 100, 100)}}

\item{stay.ethical}{numerical value, if stage 1 outcome is bigger than \code{stay.ethical} value, patient has probability of 1 of staying on the same treatment in stage 2}

\item{switch.safety}{numerical value, if stage 1 outcome is smaller than \code{switch.safety} value, patient has probability of 0 of staying on the same treatment in stage 2}

\item{wideForm}{whether to output the simulated dataset in wide form}
}
\value{
The simulated dataset with 7 variables: \code{patient id, treatment 1, stage 1 outcome, probability of stay on the same treatment, realization of staying flag (1 = stay on the same treatment, 0 = rerandomization in stage 2), treatment 2, stage 2 outcome}
}
\description{
simulate data for snSMART with continuous outcome (3 active treatments, re-randomization depends on continuous outcome at stage 1). This function generates only 1 dataset. Not to be confused with \code{\link{c_trialSim}}
}
\examples{
stage1effects<-c(50,60,70)
stage2weights<-c(.2, .8, 5)
stagecorrs<-c(.8,.3)
variance<-10
switch.safety<-NULL
stay.ethical<- NULL
#number of patients on each treatment
na<-100
nb<-100
nc<-100
n<-c(na,nb,nc)

trialData = c_trialDataGen(stage1effects, stage2weights, stagecorrs, variance, n, wideForm = TRUE)

}
\references{
Hartman, H., Tamura, R.N., Schipper, M.J. and Kidwell, K.M., 2021. Design and analysis considerations for utilizing a mapping function in a small sample,
sequential, multiple assignment, randomized trials with continuous outcomes. Statistics in Medicine, 40(2), pp.312-326.
}
\seealso{
\code{\link{c_trialSim}} \cr
\code{\link{BJSM_c}}
}
