% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_trialSim.R
\name{c_trialSim}
\alias{c_trialSim}
\title{Trial Characteristics Assessment based on Simulation (continuous snSMART design)}
\usage{
c_trialSim(
  treatInfo,
  treatCors,
  stay.ethical = NULL,
  switch.safety = NULL,
  n,
  nsim
)
}
\arguments{
\item{treatInfo}{Treatment distribution information. Each row represents one treatment, format: (mean of treatment effects, sd of treatment effects, c(priming effect on trt A,  trt B, trt C)), see example for detail}

\item{treatCors}{Treatment correlations matrix. Format example:
[[ (A,A) (A,B) (A,C)
(B,A) (B,B) (B,C)
(C,A) (C,B) (C,C) ]], (A,B) denotes the treatment correlation between first stage treatment A and second stage treatment B}

\item{stay.ethical}{numerical value, if stage 1 outcome is bigger than \code{stay.ethical} value, patient has probability of 1 of staying on the same treatment in stage 2}

\item{switch.safety}{numerical value, if stage 1 outcome is smaller than \code{switch.safety} value, patient has probability of 0 of staying on the same treatment in stage 2}

\item{n}{vector of number of patients on each treatment e.g. n <- c(100, 100, 100)}

\item{nsim}{number of simulations}
}
\value{
switch.safety: input by user; stay.ethical: input by user; stage1Eff: average simulated stage 1 treatment effect; stage2EFF: average simulated stage 2 treatment effect
stage2best: percentage of patients assigned to the best treatment in stage 2; overallbest: percentage of patients assigned to the best treatment at least once in the trial;
stage2worst: percentage of patients assigned to the worst treatment in stage 2; overallworst: percentage of patients assigned to the worst treatment at least once in the trial;
improve: percentage of patients received an improved treatment outcome in stage 2;
switchBetterOrStayedBest: percentage of patients received more effective treatments in stage 2 or stayed on the best treatment throughout the trial;
stayedSame: percentage of patients received the same treatment in stage 2 after randomization;
gotBetter: percentage of patients that received more effective treatments in stage 2;
gotWorse: percentage of patients that received less effective treatments in stage 2;
}
\description{
assess the characteristics of trial design based on simulated data
}
\examples{
treat.a<-c(70, 15, c(0,0,0))
treat.b<-c(50, 10, c(0,0,0))
treat.c<-c(60, 12, c(0,0,0))
treatInfo = rbind(treat.a, treat.b, treat.c)

treatCors<-diag(.9, 3)

switch.safety <- NULL
stay.ethical<- NULL
na<-100
nb<-100
nc<-100
n<-c(na,nb,nc)
k = c_trialSim(treatInfo, treatCors, stay.ethical, switch.safety, n, 100)

}
\references{
Hartman, H., Tamura, R.N., Schipper, M.J. and Kidwell, K.M., 2021. Design and analysis considerations for utilizing a mapping function in a small sample,
sequential, multiple assignment, randomized trials with continuous outcomes. Statistics in Medicine, 40(2), pp.312-326.
}
