% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{data_simulation}
\alias{data_simulation}
\title{Data Simulation ((snSMART with 3 active treatments and a binary outcome))}
\usage{
data_simulation(
  p_trt,
  discount_y0,
  discount_y1,
  p_1nA_2B,
  p_1nB_2A,
  p_1nC_2A,
  n
)
}
\arguments{
\item{p_trt}{vector of 3 values (first stage response rate of trt A, first stage response rate of trt B, first stage response rate of trt C). e.g. \code{p_trt = c(0.2, 0.3, 0.5)}}

\item{discount_y0}{linkage parameters, for nonresponders to treatment k in the first stage who received treatment \verb{k'} in the second stage, the second stage response rate is equal to \verb{discount_y0 * pi_IK'}}

\item{discount_y1}{linkage parameters, the second stage response rate for first stage responders is equal to \code{discount_y1 * p_IK}}

\item{p_1nA_2B}{probability for first stage non-responders to A being randomized to arm B in second stage}

\item{p_1nB_2A}{probability for first stage non-responders to B being randomized to arm A in second stage}

\item{p_1nC_2A}{probability for first stage non-responders to C being randomized to arm A in second stage}

\item{n}{vector of 3 values (first stage sample size of A, first stage sample size of B, first stage sample size of C). e.g. \code{n = c(30, 30, 30)}}
}
\value{
The simulated dataset with four columns: \code{response_stageI}, \code{treatment_stageI}, \code{response_stageII}, \code{treatment_stageII}
}
\description{
simulate data for the standard design of snSMART (3 active treatments, non-responders re-randomized; binary outcome) based on response rate of each treatment. Useful for generating large number of simulations Not to be confused with \code{\link{trial_dataset}}
}
\examples{
data = data_simulation(p_trt = c(0.2, 0.3, 0.4),
    discount_y0 = c(0.6, 0.6, 0.6), discount_y1 = c(1.5, 1.5, 1.5),
    p_1nA_2B = 0.5, p_1nB_2A = 0.5, p_1nC_2A = 0.5, n = c(30, 30, 30))

}
\references{
Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs).
Statistics in medicine, 37(26), pp.3723-3732.

Chao, Y.C., Trachtman, H., Gipson, D.S., Spino, C., Braun, T.M. and Kidwell, K.M., 2020. Dynamic treatment regimens in small n, sequential, multiple assignment, randomized trials: An application in focal segmental glomerulosclerosis. Contemporary clinical trials, 92, p.105989.
}
\seealso{
\code{\link{trial_dataset}} \cr
\code{\link{BJSM_binary}} \cr
\code{\link{JSRM_binary}} \cr
\code{\link{sample_size}}
}
