% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{data_simulation}
\alias{data_simulation}
\title{Data Simulation (standard snSMART design)}
\usage{
data_simulation(
  p_IA,
  p_IB,
  p_IC,
  discount_y0,
  discount_y1,
  p_1nA_2B,
  p_1nB_2A,
  p_1nC_2A,
  n_A,
  n_B,
  n_C
)
}
\arguments{
\item{p_IA}{first stage response rate of A}

\item{p_IB}{first stage response rate of B}

\item{p_IC}{first stage response rate of C}

\item{discount_y0}{linkage parameters, for nonresponders to treatment k in the first stage who received treatment \verb{k'} in the second stage, the second stage response rate is equal to \verb{discount_y0 * pi_IK'}}

\item{discount_y1}{linkage parameters, the second stage response rate for first stage responders is equal to \code{discount_y1 * p_IK}}

\item{p_1nA_2B}{probability for first stage non-responders to A being randomized to arm B in second stage}

\item{p_1nB_2A}{probability for first stage non-responders to B being randomized to arm A in second stage}

\item{p_1nC_2A}{probability for first stage non-responders to C being randomized to arm A in second stage}

\item{n_A}{first stage sample size of A}

\item{n_B}{first stage sample size of B}

\item{n_C}{first stage sample size of C}
}
\value{
The simulated dataset with five columns: \code{response_stageI}, \code{treatment_stageI}, \code{response_stageII}, \code{treatment_stageII}, \code{response_status_stageI} (responders will be 2 and non-responders will be 1)
}
\description{
simulate data for the standard design of snSMART (3 active treatments, non-responders re-randomized; binary outcome)
}
\examples{
data = data_simulation(p_IA = 0.2, p_IB = 0.3, p_IC = 0.4,
    discount_y0 = c(0.6, 0.6, 0.6), discount_y1 = c(1.5, 1.5, 1.5),
    p_1nA_2B = 0.5, p_1nB_2A = 0.5, p_1nC_2A = 0.5, n_A = 30, n_B = 30,
    n_C = 30)

}
\references{
Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs).
Statistics in medicine, 37(26), pp.3723-3732.

Chao, Y.C., Trachtman, H., Gipson, D.S., Spino, C., Braun, T.M. and Kidwell, K.M., 2020. Dynamic treatment regimens in small n, sequential, multiple assignment, randomized trials: An application in focal segmental glomerulosclerosis. Contemporary clinical trials, 92, p.105989.
}
