% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation_dose.R
\name{data_simulation_dose}
\alias{data_simulation_dose}
\title{Data Simulation (dose level design)}
\usage{
data_simulation_dose(
  p_IP,
  p_IL,
  p_IH,
  RESPONSE_RATE_DISCOUNT_P_Y_0,
  RESPONSE_RATE_DISCOUNT_P_Y_1,
  RESPONSE_RATE_DISCOUNT_L_Y_0,
  RESPONSE_RATE_DISCOUNT_L_Y_1,
  RESPONSE_RATE_DISCOUNT_H_Y_0,
  RESPONSE_RATE_DISCOUNT_H_Y_1,
  RAND_PROB_POS_P_TRT_H,
  RAND_PROB_POS_P_TRT_L,
  RAND_PROB_NEG_P_TRT_L,
  RAND_PROB_NEG_P_TRT_H,
  RAND_PROB_POS_L_TRT_L,
  RAND_PROB_POS_L_TRT_H,
  RAND_PROB_NEG_L_TRT_L,
  RAND_PROB_NEG_L_TRT_H,
  RAND_PROB_POS_H_TRT_L,
  RAND_PROB_POS_H_TRT_H,
  n_P,
  n_L,
  n_H
)
}
\arguments{
\item{RESPONSE_RATE_DISCOUNT_P_Y_0}{linkage parameters, for nonresponders to placebo in the first stage who received who received treatment \verb{k'} in the second stage, the second stage response rate in the second stage is equal to \verb{RESPONSE_RATE_DISCOUNT_P_Y_0 * pi_IK'}}

\item{RESPONSE_RATE_DISCOUNT_P_Y_1}{linkage parameters, for responders to placebo in the first stage who received who received treatment \verb{k'} in the second stage, the second stage response rate in the second stage is equal to \verb{RESPONSE_RATE_DISCOUNT_P_Y_1 * pi_IK'}}

\item{RESPONSE_RATE_DISCOUNT_L_Y_0}{linkage parameters, for nonresponders to low dose treatment in the first stage who received who received treatment \verb{k'} in the second stage, the second stage response rate in the second stage is equal to \verb{RESPONSE_RATE_DISCOUNT_L_Y_0 * pi_IK'}}

\item{RESPONSE_RATE_DISCOUNT_L_Y_1}{linkage parameters, for responders to low dose treatment in the first stage who received who received treatment \verb{k'} in the second stage, the second stage response rate in the second stage is equal to \verb{RESPONSE_RATE_DISCOUNT_L_Y_1 * pi_IK'}}

\item{RESPONSE_RATE_DISCOUNT_H_Y_0}{similar to \code{RESPONSE_RATE_DISCOUNT_L_Y_0}}

\item{RESPONSE_RATE_DISCOUNT_H_Y_1}{similar to \code{RESPONSE_RATE_DISCOUNT_L_Y_1}}

\item{RAND_PROB_POS_P_TRT_H}{probability for first stage responders to placebo being randomized to high dose treatment in second stage}

\item{RAND_PROB_POS_P_TRT_L}{probability for first stage responders to placebo being randomized to low dose treatment in second stage}

\item{RAND_PROB_NEG_P_TRT_L}{probability for first stage nonresponders to placebo being randomized to low dose treatment in second stage}

\item{RAND_PROB_NEG_P_TRT_H}{probability for first stage nonresponders to placebo being randomized to high dose treatment in second stage}

\item{RAND_PROB_POS_L_TRT_L}{similar to above}

\item{RAND_PROB_POS_L_TRT_H}{similar to above}

\item{RAND_PROB_NEG_L_TRT_L}{similar to above}

\item{RAND_PROB_NEG_L_TRT_H}{similar to above}

\item{RAND_PROB_POS_H_TRT_L}{similar to above}

\item{RAND_PROB_POS_H_TRT_H}{similar to above}

\item{p_IA}{first stage response rate of placebo}

\item{p_IB}{first stage response rate of low dose treatment}

\item{p_IC}{first stage response rate of high dose treatment}

\item{n_A}{first stage sample size of A}

\item{n_B}{first stage sample size of B}

\item{n_C}{first stage sample size of C}
}
\value{
The simulated dataset with five columns: response_stageI, treatment_stageI, response_stageII, treatment_stageII, response_status_stageI (responders will be two and non-responders will be one)
}
\description{
simulate data for the dose level design of snSMART (placebo, low, high dose; binary outcome)
}
\examples{
data = data_simulation_dose(p_IP = 0.15, p_IL = 0.15, p_IH = 0.15,
    RESPONSE_RATE_DISCOUNT_P_Y_0 = 0.9, RESPONSE_RATE_DISCOUNT_P_Y_1 = 1.3,
    RESPONSE_RATE_DISCOUNT_L_Y_0 = 0.8, RESPONSE_RATE_DISCOUNT_L_Y_1 = 1.2,
    RESPONSE_RATE_DISCOUNT_H_Y_0 = 0.8, RESPONSE_RATE_DISCOUNT_H_Y_1 = 1.2,
    RAND_PROB_POS_P_TRT_H = 0.5, RAND_PROB_POS_P_TRT_L = 0.5,
    RAND_PROB_NEG_P_TRT_L = 0.5, RAND_PROB_NEG_P_TRT_H = 0.5,
    RAND_PROB_POS_L_TRT_L = 0.5, RAND_PROB_POS_L_TRT_H = 0.5,
    RAND_PROB_NEG_L_TRT_L = 0.5, RAND_PROB_NEG_L_TRT_H = 0.5,
    RAND_PROB_POS_H_TRT_L = 0.5, RAND_PROB_POS_H_TRT_H = 0.5, n_P = 30,
    n_L = 30, n_H = 30)

}
\references{
Fang, F., Hochstedler, K.A., Tamura, R.N., Braun, T.M. and Kidwell, K.M., 2021. Bayesian methods to compare dose levels with placebo in a small n, sequential, multiple assignment, randomized trial.
Statistics in Medicine, 40(4), pp.963-977.
}
