% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_seq.R
\name{group_seq}
\alias{group_seq}
\title{BJSM Method for interim analysis and final analysis of group sequential trial design}
\usage{
group_seq(
  data,
  rule.type = 1,
  interim = TRUE,
  drop_threshold = 0.5,
  drop_threshold_large = 0.5,
  drop_threshold_small = 0.5,
  NUM_ARMS,
  pi_prior_dist,
  pi_prior.a,
  pi_prior.b,
  beta0_prior_dist,
  beta0_prior.a,
  beta0_prior.b,
  beta1_prior_dist,
  beta1_prior.a,
  beta1_prior.c,
  MCMC_SAMPLE,
  BURN.IN,
  n_MCMC_chain,
  ci = 0.95,
  DTR = TRUE
)
}
\arguments{
\item{data}{data should be in the same format as the simulated data generated by function \code{data_gen_group_seq_1step} and \code{data_gen_group_seq_2step}, with 8 columns:
time.1st.trt (first treatment starts time), time.1st.resp (first response time), time.2nd.trt (second treatment starts time), time.2nd.resp (second response time),
trt.1st (treatment arm for first treatment), resp.1st (response for first treatment), trt.2nd (treatment arm for second treatment), resp.2nd (response for second
treatment) data yet to be observed should be marked as "NA"}

\item{rule.type}{whether this trial is designed as a 1 step rule test  (\code{rule.type = 1}) or 2 step rule test (\code{rule.type = 2})}

\item{interim}{whether user is analyzing an interim analysis via BJSM (interim = TRUE) or an final analysis via BJSM (interim = FALSE)}

\item{drop_threshold}{only assign value to this parameter when \code{rule.type = 1} and \code{interim = TRUE} a number between 0 and 1, if the posteriors probability of a treatment having the smallest response rate is bigger than the \code{drop_threshold}, this treatment will be dropped at that update. Default value is 0.5}

\item{drop_threshold_large}{only assign value to this parameter when \code{rule.type = 2} and \code{interim = TRUE}. a number between 0 and 1. Default value is 0.5}

\item{drop_threshold_small}{only assign value to this parameter when \code{rule.type = 2} and \code{interim = TRUE}. a number between 0 and 1. Default value is 0.5. See the details section for more explanation}

\item{NUM_ARMS}{number of treatment arms}

\item{pi_prior_dist}{prior distribution for pi, user can choose from gamma, beta, pareto}

\item{pi_prior.a}{parameter a of the prior distribution for pi_1K, a vector with three values, one for each treatment. Please check the \code{Details} section for more explanation}

\item{pi_prior.b}{parameter b of the prior distribution  for pi_1K, a vector with three values, one for each treatment. Please check the \code{Details} section for more explanation}

\item{beta0_prior_dist}{prior distribution for beta0, user can choose from gamma, beta, pareto}

\item{beta0_prior.a}{parameter a of the prior distribution for linkage parameter beta0}

\item{beta0_prior.b}{parameter b of the prior distribution  for linkage parameter beta0}

\item{beta1_prior_dist}{prior distribution for beta1, user can choose from gamma, beta, pareto}

\item{beta1_prior.a}{parameter a of the prior distribution for linkage parameter beta1}

\item{beta1_prior.c}{parameter b of the prior distribution for linkage parameter beta1}

\item{MCMC_SAMPLE}{number of iterations for MCMC}

\item{BURN.IN}{number of burn-in iterations for MCMC}

\item{n_MCMC_chain}{number of MCMC chains, default to 1. If this is set to a number more than 1}

\item{ci}{coverage probability for credible intervals, default = 0.95. only assign value to this parameter when \code{interim = FALSE}.}

\item{DTR, }{if TRUE, will also return the expected response rate of dynamic treatment regimens. default = TRUE. only assign value to this parameter when \code{interim = FALSE}.}
}
\value{
if \code{interim = TRUE}, this function returns either 0 - no arm is dropped, or A/B/C - arm A/B/C is dropped
if \code{interim = FALSE}, this function returns (comments added to the code, similar to BJSM_binary.R)
}
\description{
After obtain real trial data, this function can be used to decide which arm to drop for an interim analysis or provide a full analysis for a final analysis
}
\details{
(paper provided in the reference section, section 2.2.2 Bayesian decision rules. drop_threshold_large and drop_threshold_small are corresponding to \verb{\\tau_l} and \verb{\\phi_l} respectively {need to check with Kelley})
}
\references{
Chao, Y.C., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2020. A Bayesian group sequential small n sequential multiple‐assignment randomized trial. Journal of the Royal Statistical Society: Series C (Applied Statistics), 69(3), pp.663-680.
}
