% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSRM_binary.R
\name{JSRM_binary}
\alias{JSRM_binary}
\title{JSRM}
\usage{
JSRM_binary(data, six = TRUE)
}
\arguments{
\item{data}{data format produced by the \code{trial_dataset} function}

\item{six}{if TRUE, will run the six beta model, if FALSE will run the two
beta model. Default is \code{six = TRUE}}
}
\value{
a \code{list} containing
\itemize{
\item{GEE_output}{ - original output of the GEE (geeglm) model}
\item{pi_hat}{ - estimate of response rate/treatment effect}
\item{sd_pi_hat}{ - standard error of the response rate}
\item{pi_DTR_hat}{ - DTR estimates}
\item{pi_DTR_se}{ - standard deviation of DTR estimates}
}
}
\description{
What does this function do?
}
\examples{
#data
#JSRM_binary = function(data = data, six = TRUE)
}
