% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSRM_binary.R
\name{JSRM_binary}
\alias{JSRM_binary}
\alias{summary.JSRM_binary}
\alias{print.JSRM_binary}
\title{JSRM for standard snSMART}
\usage{
JSRM_binary(data, six = TRUE, DTR = TRUE)

\method{summary}{JSRM_binary}(object, digits = 5, ...)

\method{print}{JSRM_binary}(object, digits = 5, ...)
}
\arguments{
\item{data}{data produced by \code{\link{trial_dataset}} or \code{\link{data_simulation}}}

\item{six}{if TRUE, will run the six beta model, if FALSE will run the two
beta model. Default is \code{six = TRUE}}

\item{DTR}{if TRUE, will also return the expected response rate and its standard error of dynamic treatment regimens}

\item{digits}{the number of significant digits to use when printing}
}
\value{
a \code{list} containing
\itemize{
\item{\code{GEE_output}}{ - original output of the GEE (geeglm) model}
\item{\code{pi_hat}}{ - estimate of response rate/treatment effect}
\item{\code{sd_pi_hat}}{ - standard error of the response rate}
\item{\code{pi_DTR_hat}}{ - expected response rate of dynamic treatment regimens (DTRs)}
\item{\code{pi_DTR_se}}{ - standard deviation of DTR estimates}
}
}
\description{
A joint-stage regression model (JSRM) is a frequentist modeling approach that incorporates the responses of both stages as repeated measurements for each subject.
Generalized estimating equations (GEE) are used to estimate the response rates of each treatment. The marginal response rates for each DTR can also be obtained based on the GEE results
}
\examples{
data = trial_dataset(trt = c(9, 12, 9), resp = c(3, 3, 5), trt_same_II = c(2, 2, 4),
    resp_same_II = c(2, 2, 4), trt_negA = c(3, 2), trt_negB = c(3, 2), trt_negc = c(1, 2),
    resp_negA = c(3, 1), resp_negB = c(2, 2), resp_negC = c(0, 0))

JSRM_result = JSRM_binary(data = data, six = TRUE, DTR = TRUE)

summary(JSRM_result)

}
\references{
Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs).
Statistics in medicine, 37(26), pp.3723-3732.

Chao, Y.C., Trachtman, H., Gipson, D.S., Spino, C., Braun, T.M. and Kidwell, K.M., 2020. Dynamic treatment regimens in small n, sequential, multiple assignment,
randomized trials: An application in focal segmental glomerulosclerosis. Contemporary clinical trials, 92, p.105989.
}
\seealso{
\code{\link{data_simulation}} \cr
\code{\link{trial_dataset}} \cr
\code{\link{BJSM_binary}} \cr
\code{\link{sample_size}}
}
