% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snSMART-package.R
\docType{package}
\name{snSMART-package}
\alias{snSMART}
\alias{snSMART-package}
\title{snSMART: Small N Sequential Multiple Assignment Randomized Trial Methods}
\description{
Consolidated data simulation, sample size calculation and analysis functions for several snSMART (small sample sequential, multiple assignment, randomized trial) designs under one library. See Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M. "A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs)." (2018) Statistics in medicine, 37(26), pp.3723-3732 \doi{10.1002/sim.7900}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sidiwang/snSMART}
  \item Report bugs at \url{https://github.com/sidiwang/snSMART/issues}
}

}
\author{
\strong{Maintainer}: Michael Kleinsasser \email{mkleinsa@umich.edu}

Authors:
\itemize{
  \item Sidi Wang \email{sidiwang@umich.edu}
  \item Kelley Kidwell \email{kidwell@umich.edu}
}

}
\keyword{internal}
