% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_binary.R
\name{trial_dataset}
\alias{trial_dataset}
\title{Trial dataset}
\usage{
trial_dataset(
  trtA_I,
  trtB_I,
  trtC_I,
  respA_I,
  respB_I,
  respC_I,
  trtAA_II,
  trtBB_II,
  trtCC_II,
  respA_II,
  respB_II,
  respC_II,
  trtAB_II,
  trtAC_II,
  trtBA_II,
  trtBC_II,
  trtCA_II,
  trtCB_II,
  respAB_II,
  respAC_II,
  respBA_II,
  respBC_II,
  respCA_II,
  respCB_II
)
}
\arguments{
\item{trtA_I}{number of people who receive A in stage 1}

\item{trtB_I}{number of people who receive B in stage 1}

\item{trtC_I}{number of people who receive C in stage 1}

\item{respA_I}{number of people who respond to A in stage 1, who receive same trt in stage 2}

\item{respB_I}{number of people who respond to B in stage 1, who receive same trt in stage 2}

\item{respC_I}{number of people who respond to C in stage 1, who receive same trt in stage 2}

\item{trtAA_II}{number of 1st stage responders to A who receive A in stage 2}

\item{trtBB_II}{number of 1st stage responders to B who receive B in stage 2}

\item{trtCC_II}{number of 1st stage responders to C who receive C in stage 2}

\item{respA_II}{number of 1st stage responders who respond to A again in stage 2}

\item{respB_II}{number of 1st stage responders who respond to B again in stage 2}

\item{respC_II}{number of 1st stage responders who respond to C again in stage 2}

\item{trtAB_II}{number of 1st stage non-responders to A who receive B in stage 2}

\item{trtAC_II}{number of 1st stage non-responders to A who receive C in stage 2}

\item{trtBA_II}{number of 1st stage non-responders to B who receive A in stage 2}

\item{trtBC_II}{number of 1st stage non-responders to B who receive C in stage 2}

\item{trtCA_II}{number of 1st stage non-responders to C who receive A in stage 2}

\item{trtCB_II}{number of 1st stage non-responders to C who receive B in stage 2}

\item{respAB_II}{number of 1st stage non-responders to A who respond to B in stage 2}

\item{respAC_II}{number of 1st stage non-responders to A who respond to C in stage 2}

\item{respBA_II}{number of 1st stage non-responders to B who respond to A in stage 2}

\item{respBC_II}{number of 1st stage non-responders to B who respond to C in stage 2}

\item{respCA_II}{number of 1st stage non-responders to C who respond to A in stage 2}

\item{respCB_II}{number of 1st stage non-responders to C who respond to B in stage 2}
}
\value{
a \code{matrix} of the trial dataset with 4 columns: treatment_stageI, response_stageI, treatment_stageII, response_stageII
}
\description{
Generate trial dataset with or without missing value
}
\examples{
trial_data = trial_dataset(trtA_I = 9, trtB_I = 12, trtC_I = 9, respA_I = 3,
    respB_I = 3, respC_I = 5, trtAA_II = 2, trtBB_II = 2, trtCC_II = 4,
    respA_II = 2, respB_II = 2, respC_II = 4, trtAB_II = 3, trtAC_II = 2,
    trtBA_II = 3, trtBC_II = 2, trtCA_II = 1, trtCB_II = 2, respAB_II = 3,
    respAC_II = 1, respBA_II = 2, respBC_II = 2, respCA_II = 0, respCB_II = 0)

}
\references{
Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs).
Statistics in medicine, 37(26), pp.3723-3732.
}
