% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BJSM_binary.R
\name{trial_dataset}
\alias{trial_dataset}
\title{Trial dataset generation (snSMART with 3 active treatments and a binary outcome)}
\usage{
trial_dataset(
  trt,
  resp,
  trt_same_II,
  resp_same_II,
  trt_negA,
  trt_negB,
  trt_negc,
  resp_negA,
  resp_negB,
  resp_negC
)
}
\arguments{
\item{trt}{vector of 3 values c(number of people who receive A in stage 1, number of people who receive B in stage 1, number of people who receive C in stage 1)}

\item{resp}{vector of 3 values c(number of people who respond to A in stage 1, number of people who respond to B in stage 1, number of people who respond to C in stage 1)}

\item{trt_same_II}{vector of 3 values c(number of 1st stage responders to A who receive A in stage 2, number of 1st stage responders to B who receive B in stage 2, number of 1st stage responders to C who receive C in stage 2)}

\item{resp_same_II}{vector of 3 values c(number of 1st stage responders who respond to A again in stage 2, number of 1st stage responders who respond to B again in stage 2, number of 1st stage responders who respond to C again in stage 2)}

\item{trt_negA}{vector of 2 values c(number of 1st stage non-responders to A who receive B in stage 2, number of 1st stage non-responders to A who receive C in stage 2)}

\item{trt_negB}{vector of 2 values c(number of 1st stage non-responders to B who receive A in stage 2, number of 1st stage non-responders to B who receive C in stage 2)}

\item{trt_negc}{vector of 2 values c(number of 1st stage non-responders to C who receive A in stage 2, number of 1st stage non-responders to C who receive B in stage 2)}

\item{resp_negA}{vector of 2 values c(number of 1st stage non-responders to A who respond to B in stage 2, number of 1st stage non-responders to A who respond to C in stage 2)}

\item{resp_negB}{vector of 2 values c(number of 1st stage non-responders to B who respond to A in stage 2, number of 1st stage non-responders to B who respond to C in stage 2)}

\item{resp_negC}{vector of 2 values c(number of 1st stage non-responders to C who respond to A in stage 2, number of 1st stage non-responders to C who respond to B in stage 2)}
}
\value{
a \code{matrix} of the trial dataset with 4 columns: treatment_stageI, response_stageI, treatment_stageII, response_stageII
}
\description{
Generate trial dataset based on the exact number of responders/non-responders for standard snSMART (3 active treatments, non-responders re-randomized; binary outcome) with or without missing value.
Useful for recording real snSMART trial result or generating specific simulation scenario. Not to be confused with \code{\link{data_simulation}}. Missing data in stage 2 is allowed.
}
\examples{
trial_data = trial_dataset(trt = c(9, 12, 9), resp = c(3, 3, 5), trt_same_II = c(2, 2, 4),
    resp_same_II = c(2, 2, 4), trt_negA = c(3, 2), trt_negB = c(3, 2), trt_negc = c(1, 2),
    resp_negA = c(3, 1), resp_negB = c(2, 2), resp_negC = c(0, 0))

}
\references{
Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs).
Statistics in medicine, 37(26), pp.3723-3732.
}
\seealso{
\code{\link{data_simulation}} \cr
\code{\link{BJSM_binary}} \cr
\code{\link{JSRM_binary}} \cr
\code{\link{sample_size}}
}
